Ajusta queries dos relatórios para native SQL query

+++ src/main/java/br/com/foursales/ecommerce/api/dto/TicketMedioDTO.java
@@ -1,7 +1,9 @@
 package br.com.foursales.ecommerce.api.dto;
 
+import java.math.BigDecimal;
+
 public record TicketMedioDTO(
     String cliente,
-    Double valorMedio
+    BigDecimal valorMedio
 ) {
 }
